fix issue with message character count
Ran in to an issue where a message was sent using two messages
from my carrier.  I assume this was some sort of weird bug with
the character count of the message.  Going to scale back the
message so that only 152 characters of it are shown, plus three
for the ellipsis.  This means a total of 155 characters shoud be sent
at most.